refactor: extract VehicleDetail component, validate message schemas
- Extract 302-line page.tsx into thin page (22 lines) + VehicleDetail.tsx
- Add senderId to sendMessageSchema, create markReadSchema for /read route
- Use parseBody consistently in messages routes (81 -> 63 lines)
- Remove business vehicle page from lint-file-size exemption list
- Add explicit ComponentType import for verbatimModuleSyntax safety
